LRQA-44374 Add the method logging proxy to the workspace objects
