+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/PortalWorkspace.java
@@ -14,141 +14,18 @@
 
 package com.liferay.jenkins.results.parser;
 
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-
 /**
  * @author Michael Hashimoto
  */
-public abstract class PortalWorkspace extends BaseWorkspace {
+public interface PortalWorkspace extends Workspace {
 
-	public static boolean isPortalGitHubURL(String gitHubURL) {
-		Matcher matcher = _portalGitHubURLPattern.matcher(gitHubURL);
+	public WorkspaceGitRepository getCompanionPortalWorkspaceGitRepository();
 
-		if (matcher.find()) {
-			return true;
-		}
+	public WorkspaceGitRepository getOtherPortalWorkspaceGitRepository();
 
-		return false;
-	}
+	public WorkspaceGitRepository getPluginsWorkspaceGitRepository();
 
-	protected PortalWorkspace(
-		String portalGitHubURL, String portalUpstreamBranchName) {
-
-		this(portalGitHubURL, portalUpstreamBranchName, null);
-	}
-
-	protected PortalWorkspace(
-		String portalGitHubURL, String portalUpstreamBranchName,
-		String portalBranchSHA) {
-
-		WorkspaceGitRepository workspaceGitRepository =
-			WorkspaceUtil.getWorkspaceGitRepository(
-				"portal", portalGitHubURL, portalUpstreamBranchName);
-
-		if (!(workspaceGitRepository instanceof
-				PortalWorkspaceGitRepository)) {
-
-			throw new RuntimeException("Invalid workspace Git repository");
-		}
-
-		_primaryPortalWorkspaceGitRepository =
-			(PortalWorkspaceGitRepository)workspaceGitRepository;
-
-		if (portalBranchSHA != null) {
-			_primaryPortalWorkspaceGitRepository.setBranchSHA(portalBranchSHA);
-		}
-
-		_primaryPortalWorkspaceGitRepository.setUp();
-
-		_companionPortalWorkspaceGitRepository =
-			WorkspaceUtil.getDependencyWorkspaceGitRepository(
-				CompanionPortalWorkspaceGitRepository.TYPE,
-				_primaryPortalWorkspaceGitRepository);
-
-		if (_companionPortalWorkspaceGitRepository != null) {
-			_companionPortalWorkspaceGitRepository.setUp();
-		}
-
-		_otherPortalWorkspaceGitRepository =
-			WorkspaceUtil.getDependencyWorkspaceGitRepository(
-				OtherPortalWorkspaceGitRepository.TYPE,
-				_primaryPortalWorkspaceGitRepository);
-
-		_pluginsWorkspaceGitRepository =
-			WorkspaceUtil.getDependencyWorkspaceGitRepository(
-				PortalPluginsWorkspaceGitRepository.TYPE,
-				_primaryPortalWorkspaceGitRepository);
-	}
-
-	protected WorkspaceGitRepository getOtherPortalWorkspaceGitRepository() {
-		return _otherPortalWorkspaceGitRepository;
-	}
-
-	protected WorkspaceGitRepository getPluginsWorkspaceGitRepository() {
-		return _pluginsWorkspaceGitRepository;
-	}
-
-	protected PortalWorkspaceGitRepository
-		getPrimaryPortalWorkspaceGitRepository() {
-
-		return _primaryPortalWorkspaceGitRepository;
-	}
-
-	@Override
-	protected void setUpWorkspaceGitRepositories() {
-		setUpJenkinsWorkspaceGitRepository();
-
-		_primaryPortalWorkspaceGitRepository.setUp();
-
-		if (_companionPortalWorkspaceGitRepository != null) {
-			_companionPortalWorkspaceGitRepository.setUp();
-		}
-
-		if (_otherPortalWorkspaceGitRepository != null) {
-			_otherPortalWorkspaceGitRepository.setUp();
-		}
-
-		_pluginsWorkspaceGitRepository.setUp();
-	}
-
-	@Override
-	protected void setWorkspaceGitRepositoryJobProperties(Job job) {
-		_primaryPortalWorkspaceGitRepository.setPortalJobProperties(job);
-	}
-
-	@Override
-	protected void tearDownWorkspaceGitRepositories() {
-		tearDownJenkinsWorkspaceGitRepository();
-
-		_primaryPortalWorkspaceGitRepository.tearDown();
-
-		if (_companionPortalWorkspaceGitRepository != null) {
-			_companionPortalWorkspaceGitRepository.tearDown();
-		}
-
-		if (_otherPortalWorkspaceGitRepository != null) {
-			_otherPortalWorkspaceGitRepository.tearDown();
-		}
-
-		_pluginsWorkspaceGitRepository.tearDown();
-	}
-
-	@Override
-	protected void writeWorkspaceGitRepositoryPropertiesFiles() {
-		_primaryPortalWorkspaceGitRepository.writePropertiesFiles();
-
-		_pluginsWorkspaceGitRepository.writePropertiesFiles();
-	}
-
-	private static final Pattern _portalGitHubURLPattern = Pattern.compile(
-		"https://github.com/[^/]+/(?<gitRepositoryName>" +
-			"liferay-portal(-ee)?)/.*");
-
-	private final WorkspaceGitRepository _companionPortalWorkspaceGitRepository;
-	private final WorkspaceGitRepository _otherPortalWorkspaceGitRepository;
-	private final WorkspaceGitRepository _pluginsWorkspaceGitRepository;
-	private final PortalWorkspaceGitRepository
-		_primaryPortalWorkspaceGitRepository;
+	public PortalWorkspaceGitRepository
+		getPrimaryPortalWorkspaceGitRepository();
 
 }
